ARM: shmobile: r8a7779: Set the SGX clock to clks
Whilst it's possible to change the SGX clock speed, we set it to
the maximum (i.e. clks).

The register used to control this, is under the influence of the
SYSC SGX power domain. You can't set it unless it's powered up.
For this reason, the code to has been added to the PM code, not the
clock code.