mmc: sh-mmcif: Fix MMC_WRITE_MULTIPLE_BLOCK completion IRQ handling
Upon completion of a MMC_WRITE_MULTIPLE_BLOCK command MMCIF issues an IRQ
with the BUFREN or BUFWEN bits set and often with one or several of CMD12
bits set.

If those interrupts are not acknowledged, an additional interrupt can be
produced and will be delivered later, possibly when the transaction has
already been completed. To prevent this from happening, CMD12 completion
interrupt sources have to be cleared too upon reception of a BUFREN or
BUFWEN IRQ.

This is similar to commit 7a7eb328; this change adds the same handling for
BUFREN and BUFWEN interrupts.